+++ tests/Tests/PeachApiClientTests.cs
@@ -5,7 +5,9 @@ using Xunit.Abstractions;
 
 public class PeachApiClientTests(ITestOutputHelper output)
 {
-    [Fact]
+    const int DEFAULT_DELAY_MS = 1000;
+
+    [Fact, Delay(DEFAULT_DELAY_MS)]
     public async Task Search_all_offer_types() => await SeachOffersAndAssertAsync(new OfferFilter {
         Type = OfferTypeFilter.All,
         Amount = [100_000, 1_000_000],
@@ -14,7 +16,7 @@ public class PeachApiClientTests(ITestOutputHelper output)
         MinReputation = 0.5
     });
 
-    [Fact]
+    [Fact, Delay(DEFAULT_DELAY_MS)]
     public async Task Search_ask_offer_types() => await SeachOffersAndAssertAsync(new OfferFilter {
         Type = OfferTypeFilter.Ask,
         Amount = [100_000, 1_000_000],
@@ -24,7 +26,7 @@ public class PeachApiClientTests(ITestOutputHelper output)
     },
     assert: offers => Assert.All(offers, o => Assert.Equal(OfferType.Ask, o.Type)));
 
-    [Fact]
+    [Fact, Delay(DEFAULT_DELAY_MS)]
     public async Task Search_bid_offer_types() => await SeachOffersAndAssertAsync(new OfferFilter {
         Type = OfferTypeFilter.Bid,
         Amount = [100_000, 1_000_000],
@@ -34,12 +36,12 @@ public class PeachApiClientTests(ITestOutputHelper output)
     },
     assert: offers => Assert.All(offers, o => Assert.Equal(OfferType.Bid, o.Type)));
 
-    [Fact]
+    [Fact, Delay(DEFAULT_DELAY_MS)]
     public async Task Search_offers_with_only_mandatory_filters() => await SeachOffersAndAssertAsync(new OfferFilter {
         Type = OfferTypeFilter.All
     });
 
-    [Fact]
+    [Fact, Delay(DEFAULT_DELAY_MS)]
     public async Task Search_offers_with_pagination() => await SeachOffersAndAssertAsync(new OfferFilter {
         Type = OfferTypeFilter.All,
     },
@@ -47,7 +49,7 @@ public class PeachApiClientTests(ITestOutputHelper output)
     assert: offers => Assert.Single(offers),
     failOnEmpty: true);
 
-    [Fact]
+    [Fact, Delay(DEFAULT_DELAY_MS)]
     public async Task Search_offers_with_sort() => await SeachOffersAndAssertAsync(new OfferFilter {
        Type = OfferTypeFilter.Ask, },
        pagination: new OfferPagination(0, 2),
@@ -59,7 +61,6 @@ public class PeachApiClientTests(ITestOutputHelper output)
        },
        failOnEmpty: true);
     
-
     private async Task SeachOffersAndAssertAsync(OfferFilter filter, OfferPagination? pagination = null,
         OfferSortBy? sort = null,
         Action<List<Offer>>? assert = null, bool failOnEmpty = false)